PaintWorklet: Move paintWorklet from 'window' to 'CSS'
Before this CL, |paintWorklet| attribute is defined on Window interface, but
that should be defined on CSS interface[1].

According to the spec discussion[2], "CSS" will be changed from "interface" to
"namespace", but "namespace" is not supported yet in Blink. As a stopgap, this
CL changes "paintWorklet" to a static attribute to emulate an
attribute-on-namespace and moves it to "CSS" interface.

[1] https://drafts.css-houdini.org/css-paint-api-1/#dom-css-paintworklet
[2] https://github.com/w3c/css-houdini-drafts/issues/410

Bug: 719303
Change-Id: I01f371f0d4f9a193352d1ede1fe28632d52cf312
Reviewed-on: https://chromium-review.googlesource.com/630899
WPT-Export-Revision: c3dcf02b902dbdbf35d3f700d897b71e1dbc4042